+++ tests/tinytest/test_case_when.R
@@ -59,3 +59,15 @@ expect_equal(
     pull(y),
   c("foo", "foo", NA, "foo", NA)
 )
+
+# evaluation of external objects works
+
+foo <<- "a"
+foo2 <<- "b"
+
+expect_equal(
+  test |>
+    mutate(y = case_when(x1 %in% foo ~ "foo", x1 %in% foo2 ~ "foo2", .default = x1)) |>
+    pull(y),
+  c("foo", "foo", "foo2", "foo", "c")
+)