+++ tests/tinytest/test_ifelse.R
@@ -53,3 +53,14 @@ expect_equal(
     pull(y),
   c("hello", "hello", "b", "hello", "c")
 )
+
+# evaluation of external objects works
+
+foo <<- "a"
+
+expect_equal(
+  test |>
+    mutate(y = ifelse(x1 %in% foo, x3, x1)) |>
+    pull(y),
+  c("hello", "hello", "b", "hello", "c")
+)